+++ backend/src/routes/task.ts
@@ -7,9 +7,14 @@ import {
   deleteTask,
 } from "../repositories/task";
 import { getTaskAndVerify, updateTask } from "../services/task";
-import { getProject } from "../repositories/project";
+import { getProjectFromDB } from "../repositories/project";
+import { getSectionById } from "../repositories/section";
 import { getUserIdFromRequest } from "../middleware/auth";
-import { BadRequestError, NotFoundError } from "../errors/errors";
+import {
+  BadRequestError,
+  NotFoundError,
+  ConflictError,
+} from "../errors/errors";
 import {
   type CreateTaskQuery,
   CreateTaskQuerySchema,
@@ -51,12 +56,30 @@ function addTaskCreateRoute(fastify: FastifyInstance) {
     },
     async function handler(request, reply) {
       const userId = getUserIdFromRequest(request);
-      const project = await getProject(request.body.project_id);
+      const project = await getProjectFromDB(request.body.project_id);
 
       if (!project || project.creator_id !== userId) {
         throw new BadRequestError(
           "Project id either does not exist or you don't have access to it"
         );
+      }
+
+      if (project.is_archived) {
+        throw new ConflictError("Cannot create tasks in archived projects");
+      }
+
+      if (request.body.section_id) {
+        const section = await getSectionById(request.body.section_id);
+
+        if (!section || section.creator_id !== userId) {
+          throw new BadRequestError(
+            "Section id either does not exist or you don't have access to it"
+          );
+        }
+
+        if (section.is_archived) {
+          throw new ConflictError("Cannot create tasks in archived sections");
+        }
       }
 
       const task = await createTaskInDB({
@@ -87,7 +110,7 @@ function addTasksGetRoute(fastify: FastifyInstance) {
     async function handler(request, reply) {
       const projectId = request.query.project_id;
       const userId = getUserIdFromRequest(request);
-      const project = await getProject(projectId);
+      const project = await getProjectFromDB(projectId);
 
       if (!project || project.creator_id !== userId) {
         throw new BadRequestError(
